feat: added validation of user input and implemented HikariCP

+++ src/main/java/roadmap/model/ExchangeRateResponse.java
@@ -3,6 +3,8 @@ package roadmap.model;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import roadmap.model.dto.CurrencyDto;
 
+import java.math.BigDecimal;
+
 @JsonPropertyOrder({"id", "baseCurrency", "targetCurrency", "rate"})
-public record ExchangeRateResponse(Long id, CurrencyDto baseCurrency, CurrencyDto targetCurrency, Double rate) {
+public record ExchangeRateResponse(Long id, CurrencyDto baseCurrency, CurrencyDto targetCurrency, BigDecimal rate) {
 }
